+++ packages/fuselage/src/components/Message/MessageReactions/MessageReactions.stories.tsx
@@ -1,4 +1,5 @@
 import type { Meta } from '@storybook/react';
+import type { ComponentType } from 'react';
 
 import { BasicMessageTemplate } from '../helpers';
 
@@ -10,8 +11,8 @@ export default {
   title: 'Message/MessageReactions',
   component: MessageReactions,
   subcomponents: {
-    MessageReaction,
-    MessageReactionAction,
+    MessageReaction: MessageReaction as ComponentType<any>,
+    MessageReactionAction: MessageReactionAction as ComponentType<any>,
   },
 } satisfies Meta<typeof MessageReactions>;