+++ packages/fuselage/src/components/Table/Table.stories.tsx
@@ -1,4 +1,5 @@
 import type { Meta, StoryFn } from '@storybook/react';
+import type { ComponentType } from 'react';
 
 import { CheckBox } from '../CheckBox';
 
@@ -17,13 +18,13 @@ export default {
   title: 'Data Display/Table',
   component: Table,
   subcomponents: {
-    TableHead,
-    TableRow,
-    TableCell,
-    TableBody,
-    TableSelection,
-    TableSelectionButton,
-    TableSelectionButtonGroup,
+    TableHead: TableHead as ComponentType<any>,
+    TableRow: TableRow as ComponentType<any>,
+    TableCell: TableCell as ComponentType<any>,
+    TableBody: TableBody as ComponentType<any>,
+    TableSelection: TableSelection as ComponentType<any>,
+    TableSelectionButton: TableSelectionButton as ComponentType<any>,
+    TableSelectionButtonGroup: TableSelectionButtonGroup as ComponentType<any>,
   },
 } satisfies Meta<typeof Table>;