+++ packages/layout/src/List/List.tsx
@@ -1,12 +1,6 @@
-import type { Box } from '@rocket.chat/fuselage';
 import { Margins } from '@rocket.chat/fuselage';
 import colors from '@rocket.chat/fuselage-tokens/colors.json';
-import type {
-  ReactElement,
-  ReactNode,
-  ComponentProps,
-  CSSProperties,
-} from 'react';
+import type { ReactNode, ComponentProps, CSSProperties } from 'react';
 
 import { ListComponent } from './List.styles';
 import ListItem from './ListItem';
@@ -21,9 +15,9 @@ const List = ({
   children: ReactNode;
   spacing?: ComponentProps<typeof Margins>['block'];
   listStyleType?: CSSProperties['listStyleType'];
-  color?: ComponentProps<typeof Box>['color'];
+  color?: CSSProperties['color'];
   icon?: string;
-}): ReactElement => (
+}) => (
   <ListComponent icon={icon} listStyleType={listStyleType} color={color}>
     <Margins block={spacing}>{children}</Margins>
   </ListComponent>