+++ packages/onboarding-ui/src/forms/CreateFirstMemberForm/CreateFirstMemberForm.spec.tsx
@@ -1,4 +1,4 @@
-import ReactDOM from 'react-dom';
+import { render } from '@testing-library/react';
 
 import CreateFirstMemberForm from './CreateFirstMemberForm';
 
@@ -18,7 +18,5 @@ const props = {
 };
 
 it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<CreateFirstMemberForm {...props} />, div);
-  ReactDOM.unmountComponentAtNode(div);
+  render(<CreateFirstMemberForm {...props} />);
 });