+++ packages/onboarding-ui/src/forms/TotpForm/TotpForm.spec.tsx
@@ -1,16 +1,13 @@
-import ReactDOM from 'react-dom';
+import { render } from '@testing-library/react';
 
 import TotpForm from './TotpForm';
 
 it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(
+  render(
     <TotpForm
       onChangeTotpForm={() => undefined}
       isBackupCode={false}
       onSubmit={() => undefined}
     />,
-    div,
   );
-  ReactDOM.unmountComponentAtNode(div);
 });